Separate routes for notes and versions
