+++ src/routes/__init__.py
@@ -1,1 +1,2 @@
 from routes.notes import router as note_router
+from routes.versions import router as version_router
