+++ src/routes/notes.py
@@ -11,15 +11,13 @@ from schemas import (
     NoteDetailResponseSchema,
     NoteCreateRequestSchema,
     NoteUpdateRequestSchema,
-    VersionListResponseSchema,
-    VersionDetailResponseSchema,
 )
 
 router = APIRouter()
 
 
 @router.get("/notes/", response_model=NoteListResponseSchema)
-async def get_notes(
+async def get_note_list(
     page: int = Query(1, ge=1),
     per_page: int = Query(10, ge=1, le=20),
     db: AsyncSession = Depends(get_db),
@@ -66,7 +64,7 @@ async def get_notes(
 
 
 @router.get("/notes/{note_id}/", response_model=NoteDetailResponseSchema)
-async def get_note(note_id: int, db: AsyncSession = Depends(get_db)):
+async def retrieve_note(note_id: int, db: AsyncSession = Depends(get_db)):
     """
     Retrieve a single note by ID.
     Returns the note with the given ID.
@@ -114,7 +112,7 @@ async def update_note(
     Creates a new version with the previous note content and updates the note with new content.
     Returns the updated note with its versions.
     """
-    note = await get_note(note_id, db)
+    note = await retrieve_note(note_id, db)
 
     # Check the latest version of the note
     latest_version_result = await db.execute(
@@ -146,84 +144,8 @@ async def delete_note(note_id: int, db: AsyncSession = Depends(get_db)):
     """
     Delete a note by ID.
     """
-    note = await get_note(note_id, db)
+    note = await retrieve_note(note_id, db)
 
     await db.delete(note)
     await db.commit()
     return {"message": "Note deleted successfully."}
-
-
-@router.get("/notes/{note_id}/versions/", response_model=VersionListResponseSchema)
-async def get_versions(
-    note_id: int,
-    page: int = Query(1, ge=1),
-    per_page: int = Query(10, ge=1, le=100),
-    db: AsyncSession = Depends(get_db),
-):
-    """
-    Retrieve a paginated list of versions by note ID.
-    Returns a list of versions with pagination metadata.
-    """
-    note = await get_note(note_id, db)
-
-    versions = note.versions
-    total_versions = len(versions)
-    total_pages = (total_versions + per_page - 1) // per_page
-
-    offset = (page - 1) * per_page
-    paginated_versions = versions[offset : offset + per_page]
-
-    return {
-        "versions": paginated_versions,
-        "prev_page": (
-            f"/notes/{note_id}/versions/?page={page - 1}&per_page={per_page}"
-            if page > 1
-            else None
-        ),
-        "next_page": (
-            f"/notes/{note_id}/versions/?page={page + 1}&per_page={per_page}"
-            if page < total_pages
-            else None
-        ),
-        "total_pages": total_pages,
-        "total_items": total_versions,
-    }
-
-
-@router.get(
-    "/notes/{note_id}/versions/{version_id}", response_model=VersionDetailResponseSchema
-)
-async def get_version(
-    note_id: int, version_id: int, db: AsyncSession = Depends(get_db)
-):
-    """
-    Retrieve a single version of a note by note ID and version ID.
-    Returns the version.
-    """
-
-    result = await db.execute(
-        select(VersionModel)
-        .where(VersionModel.note_id == note_id)
-        .where(VersionModel.version == version_id)
-    )
-
-    version = result.scalar_one_or_none()
-    if not version:
-        raise HTTPException(
-            status_code=404, detail="Version with the given ID was not found."
-        )
-    return version
-
-
-@router.delete("/notes/{note_id}/versions/{version_id}")
-async def delete_version(
-    note_id: int, version_id: int, db: AsyncSession = Depends(get_db)
-):
-    """
-    Delete a version of a note by note ID and version ID.
-    """
-    version = await get_version(note_id, version_id, db)
-
-    await db.delete(version)
-    await db.commit()
-    return {"message": "Version deleted successfully."}
